49 endpoint save participant session data (#52)
removed study_factor and study_task
updated queries to reflect study_factor and study_task removal
fixed duration input handling for python
added tables for participants and their session data
added better error handling for easier messaging of when the server fails
made a testing vue for uploading CSV file and reading server response
tinkered with the server so developers may access the CSV directory on their own user accounts. I made sure that this directory is open only to our user accounts and not any and all users on the server
THINGS TO KNOW:

The endpoint works, but its inputs are mostly hardcoded. (task_id, factor_id, etc) so this makes the code in a DEMO state NOT Final state
Everyone needs to update their env with the file path. Check your discords I'll send it there
Participant session data CSV is stored in the filesystem of the server because a filesystem is much better for this than using a relational database. I'm talking significant benefits in terms of storage, access, and not bloating the database with 27,000,000 (i did the math) entries for a standard study completion with mouse tracking. The database session_data_instance has a field which is the path to where its corresponding CSV file is. The flask endpoint can build / traverse directories which is sick. The hierarchy of the file system directories and using ids as the naming convention should make updating/searching mega quick
I have not yet updated my drop_tables query to also delete the CSVs. Once we deploy, this is a big no go. However, for development purposes until then, both the sub-directories and tables must be deleted together